budgetagent/tests/test_import_sweden_csv.py
"""
Test import of Swedish bank CSV files.

Testsuite för import_bank_data-modulen som hanterar inläsning av svenska bankutdrag
från CSV, Excel och JSON. Täcker fil-inläsning, format-detektering och
kolumn-normalisering.
"""

import pytest
import pandas as pd
import yaml
from pathlib import Path

# Funktioner att testa (importeras när de är implementerade)
# from budgetagent.modules.import_bank_data import (
#     load_file,
#     detect_format,
#     normalize_columns,
# )


class TestLoadFile:
    """Tester för load_file-funktionen."""

    def test_load_csv_file(self):
        """Test att läsa in en CSV-fil."""
        # TODO: Implementera test när load_file är implementerad
        pass

    def test_load_excel_file(self):
        """Test att läsa in en Excel-fil."""
        # TODO: Implementera test när load_file är implementerad
        pass

    def test_load_nonexistent_file(self):
        """Edge case: Försök läsa fil som inte finns."""
        # TODO: Implementera test för felhantering
        pass

    def test_load_empty_file(self):
        """Edge case: Tom fil."""
        # TODO: Implementera test för tom fil
        pass

    def test_load_malformed_csv(self):
        """Edge case: Felaktigt formaterad CSV."""
        # TODO: Implementera test för felaktig formatering
        pass


class TestDetectFormat:
    """Tester för detect_format-funktionen."""

    def test_detect_swedbank_format(self):
        """Test att detektera Swedbank-format."""
        # Exempel: DataFrame med Swedbank-kolumner
        # TODO: Implementera test när detect_format är implementerad
        pass

    def test_detect_seb_format(self):
        """Test att detektera SEB-format."""
        # TODO: Implementera test när detect_format är implementerad
        pass

    def test_detect_revolut_format(self):
        """Test att detektera Revolut-format."""
        # TODO: Implementera test när detect_format är implementerad
        pass

    def test_detect_unknown_format(self):
        """Edge case: Okänt bankformat."""
        # TODO: Implementera test för okänt format
        pass

    def test_detect_empty_dataframe(self):
        """Edge case: Tom DataFrame."""
        empty_df = pd.DataFrame()
        # TODO: Implementera test för tom DataFrame
        pass


class TestNormalizeColumns:
    """Tester för normalize_columns-funktionen."""

    def test_normalize_swedbank_columns(self):
        """Test att normalisera Swedbank-kolumner till standardformat."""
        # Exempel på standardisering till: date, amount, description, currency
        # TODO: Implementera test när normalize_columns är implementerad
        pass

    def test_normalize_seb_columns(self):
        """Test att normalisera SEB-kolumner till standardformat."""
        # TODO: Implementera test när normalize_columns är implementerad
        pass

    def test_normalize_with_missing_columns(self):
        """Edge case: DataFrame med saknade kolumner."""
        # TODO: Implementera test för ofullständig data
        pass

    def test_normalize_with_invalid_format(self):
        """Edge case: Felaktigt format-argument."""
        # TODO: Implementera test för ogiltigt format
        pass


class TestYAMLValidation:
    """Tester för YAML-konfigurationsvalidering."""

    def test_settings_panel_yaml_exists(self):
        """Validera att settings_panel.yaml existerar och är giltig."""
        config_path = Path(__file__).parent.parent / "config" / "settings_panel.yaml"
        assert config_path.exists(), "settings_panel.yaml saknas"

        with open(config_path, "r", encoding="utf-8") as f:
            config = yaml.safe_load(f)

        assert config is not None, "YAML-filen är tom eller ogiltig"
        assert "settings_panel" in config, "settings_panel-nyckel saknas i YAML"

    def test_import_format_configuration(self):
        """Validera import_format-konfiguration i settings_panel.yaml."""
        config_path = Path(__file__).parent.parent / "config" / "settings_panel.yaml"

        with open(config_path, "r", encoding="utf-8") as f:
            config = yaml.safe_load(f)

        # Kontrollera att import_format finns och har rätt struktur
        if "settings_panel" in config and "import_format" in config["settings_panel"]:
            import_format = config["settings_panel"]["import_format"]
            assert "type" in import_format, "type saknas i import_format"
            assert "options" in import_format, "options saknas i import_format"
            assert isinstance(
                import_format["options"], list
            ), "options ska vara en lista"


class TestIntegration:
    """Integrationstester för hela import-flödet."""

    def test_full_import_workflow(self):
        """Test av hela import-flödet: load -> detect -> normalize."""
        # TODO: Implementera end-to-end test
        pass

    def test_import_with_empty_input(self):
        """Edge case: Tomt importflöde."""
        # TODO: Implementera test för tomt flöde
        pass
