requirements.txt
# BudgetAgent Dependencies
# Install with: pip install -r requirements.txt

# Core dependencies
dash>=2.0.0
plotly>=5.0.0
pandas>=1.3.0
pyyaml>=6.0

# Testing
pytest>=7.0.0
yamllint>=1.26.0
